Update cuda stream manager logic
