+++ tensorrt_llm/_torch/auto_deploy/custom_ops/attention_interface.py
@@ -615,15 +615,7 @@ class SequenceInfo:
 
     @nvtx_range("ad_get_unique_value")
     def _get_unique_value(self, occupied: Set[int], max_val: int) -> int:
-        """Get un unoccupied value from the range indicated by max_val.
-
-        In addition, this function performs a sanity check to ensure that no value in the occupied
-        set is out of bounds.
-        """
-        # Validate without materializing the full range set
-        out_of_range = [v for v in occupied if v < 0 or v >= max_val]
-        assert not out_of_range, f"Out of range values: {out_of_range}"
-
+        """Get un unoccupied value from the range indicated by max_val."""
         # Return the smallest free value; fall back to 0 if none
         for candidate in range(max_val):
             if candidate not in occupied:
